+++ lib/additionals/patches/auto_completes_controller_patch.rb
@@ -33,13 +33,13 @@ module Additionals
         end
 
         def assignee
-          scope = @project ? @project.principals : Principal.assignable
+          scope = @project ? @project.principals.visible : Principal.assignable
 
           render_grouped_users_with_select2 scope, search_term: @search_term
         end
 
         def grouped_users
-          scope = @project ? @project.users : User.visible
+          scope = @project ? @project.users.visible : User.visible
           scope = scope.where.not id: params[:user_id] if params[:user_id].present?
           scope = scope.active if RedminePluginKit.true? params[:active_only]
 
@@ -53,7 +53,7 @@ module Additionals
 
         # user and groups
         def grouped_principals
-          scope = @project ? @project.principals : Principal.assignable
+          scope = @project ? @project.principals.visible : Principal.assignable
 
           render_params = { search_term: @search_term,
                             with_me: RedminePluginKit.true?(params[:with_me]) }