+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Changelog
 - Switch from gemoji to tanuki_emoji
 - Chart.js 3 support
 - Add Chart.js matrix plugin
+- Fix scope of users for select2 in projects
 
 3.0.5.2
 +++++++